Move network calls and transformation to services folder
