+++ app/src/TradePage/OpenPositions/types.ts
@@ -1,6 +1,7 @@
+import type { ReactNode } from 'react';
 import type { OpenPosition } from '../common/types';
 
 export type Props = {
   openPositions: OpenPosition[];
-  lastUpdated: JSX.Element;
+  lastUpdated: ReactNode;
 };
