+++ app/src/TradePage/Overview/types.ts
@@ -1,8 +1,9 @@
+import type { ReactNode } from 'react';
 import type { Trade, DepositsWithdrawal, EquitySummaryInBase } from '../common/types';
 
 export type Props = {
   trades: Trade[];
   depositsWithdrawals: DepositsWithdrawal[];
   equitySummaryInBase: EquitySummaryInBase[];
-  lastUpdated: JSX.Element;
+  lastUpdated: ReactNode;
 };
