+++ app/src/TradePage/TradeHistory/types.ts
@@ -1,7 +1,8 @@
+import type { ReactNode } from 'react';
 import type { Trade } from '../common/types';
 
 export type Props = {
   trades: Trade[];
-  lastUpdated: JSX.Element;
+  lastUpdated: ReactNode;
   timezone: string;
 };
