+++ app/src/TradePage/services/Last365CalendarDays/types.ts
@@ -1,0 +1,3 @@
+import { Last365CalendarDays } from '../types';
+
+export type GetLast365CalendarDaysResponseType = Last365CalendarDays | null;
