+++ app/src/TradePage/services/index.ts
@@ -1,0 +1,2 @@
+export { getLast365CalendarDays } from './Last365CalendarDays';
+export { transformData } from './transformer';
