+++ app/src/TradePage/services/transformer/types.ts
@@ -1,0 +1,5 @@
+import { Last365CalendarDays } from '../types';
+
+export type TransformDataProps = {
+  last365CalendarDays: Last365CalendarDays;
+};
